added ketod schema variations

+++ src/utilities/tokenizer_utilities.py
@@ -6,7 +6,7 @@ class TokenizerUtilities:
     @staticmethod
     def tokenize(tokenizer: AutoTokenizer, text: str, max_len: int = None):
         tokens = tokenizer.encode(text, return_tensors="pt", max_length=max_len)
-        return tokens.to(dtype=torch.int16)[0]
+        return tokens.to(dtype=torch.int32)[0]
 
     @staticmethod
     def tokenize_with_pad(tokenizer: AutoTokenizer, text: str, max_len: int = None):
